show granule only if selected date is within granule date range
