+++ web/js/components/timeline/timeline-coverage/timeline-coverage.js
@@ -19,97 +19,6 @@ import Switch from '../../util/switch';
 import LayerCoverageInfoModal from './info-modal';
 import CoverageItemList from './coverage-item-list';
 
-function makeTime(date) {
-  return new Date(date).getTime();
-}
-
-function mergeSortedGranuleDateRanges(granules) {
-  return granules.reduce((acc, [start, end]) => {
-    if (!acc.length) return [[start, end]];
-    const startTime = makeTime(start);
-    const endTime = makeTime(end);
-    const lastRangeEndTime = makeTime(acc.at(-1)[1]);
-    const lastRangeStartTime = makeTime(acc.at(-1)[0]);
-    if ((startTime >= lastRangeStartTime && startTime <= lastRangeEndTime) && (endTime >= lastRangeStartTime && endTime <= lastRangeEndTime)) { // within current range, ignore
-      return acc;
-    }
-    if (startTime > lastRangeEndTime) { // discontinuous, add new range
-      return [...acc, [start, end]];
-    }
-    if (startTime <= lastRangeEndTime && endTime > lastRangeEndTime) { // intersects current range, merge
-      return acc.with(-1, [acc.at(-1)[0], end]);
-    }
-    return acc;
-  }, []);
-}
-
-async function requestGranules(params) {
-  const {
-    shortName,
-    extent,
-    startDate,
-    endDate,
-  } = params;
-  const granules = [];
-  let hits = Infinity;
-  let searchAfter = false;
-  const url = `https://cmr.earthdata.nasa.gov/search/granules.json?shortName=${shortName}&bounding_box=${extent.join(',')}&temporal=${startDate}/${endDate}&sort_key=start_date&pageSize=2000`;
-  /* eslint-disable no-await-in-loop */
-  do { // run the query at least once
-    const headers = searchAfter ? { 'Cmr-Search-After': searchAfter, 'Client-Id': 'Worldview' } : { 'Client-Id': 'Worldview' };
-    const res = await fetch(url, { headers });
-    searchAfter = res.headers.get('Cmr-Search-After');
-    hits = parseInt(res.headers.get('Cmr-Hits'), 10);
-    const data = await res.json();
-    granules.push(...data.feed.entry);
-  } while (searchAfter || hits > granules.length); // searchAfter will not be present if there are no more results https://cmr.earthdata.nasa.gov/search/site/docs/search/api.html#search-after
-
-  return granules;
-}
-
-async function getLayerGranuleRanges(layer) {
-  const extent = [-180, -90, 180, 90];
-  const startDate = new Date(layer.startDate).toISOString();
-  const endDate = layer.endDate ? new Date(layer.endDate).toISOString() : new Date().toISOString();
-  const shortName = layer.conceptIds?.[0]?.shortName;
-  const nrtParams = {
-    shortName,
-    extent,
-    startDate,
-    endDate,
-  };
-  const nrtGranules = await requestGranules(nrtParams);
-  let nonNRTGranules = [];
-  if (shortName.includes('_NRT')) { // if NRT, also get non-NRT granules
-    const nonNRTShortName = shortName.replace('_NRT', '');
-    const nonNRTParams = {
-      shortName: nonNRTShortName,
-      extent,
-      startDate,
-      endDate,
-    };
-    nonNRTGranules = await requestGranules(nonNRTParams);
-  }
-  const granules = [...nonNRTGranules, ...nrtGranules];
-  const granuleDateRanges = granules.map(({ time_start: timeStart, time_end: timeEnd }) => [timeStart, timeEnd]);
-  const mergedGranuleDateRanges = mergeSortedGranuleDateRanges(granuleDateRanges); // merge overlapping granule ranges to simplify rendering
-
-  return mergedGranuleDateRanges;
-}
-
-async function mapGranulesToLayers(layers) {
-  const promises = layers.map(async (layer) => {
-    if (!layer.cmrAvailability) return layer;
-
-    const ranges = await getLayerGranuleRanges(layer);
-
-    return { ...layer, granules: ranges };
-  });
-  const cmrLayers = await Promise.all(promises);
-
-  return cmrLayers;
-}
-
 /*
  * Timeline Layer Coverage Panel for temporal coverage.
  *
@@ -120,7 +29,6 @@ class TimelineLayerCoveragePanel extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      cmrLayers: [],
       activeLayers: [],
       shouldIncludeHiddenLayers: false,
     };
@@ -226,8 +134,8 @@ class TimelineLayerCoveragePanel extends Component {
       futureTime, ongoing,
     } = layer;
 
-    if (layer.granules?.length) {
-      return layer.granules.map(([startDate, endDate]) => {
+    if (layer.granuleDateRanges?.length) {
+      return layer.granuleDateRanges.map(([startDate, endDate]) => {
         const { gridWidth } = timeScaleOptions[timeScale].timeAxis;
         const axisFrontDate = new Date(frontDate).getTime();
         const axisBackDate = new Date(backDate).getTime();
@@ -357,11 +265,9 @@ class TimelineLayerCoveragePanel extends Component {
   // eslint-disable-next-line react/destructuring-assignment
   addMatchingCoverageToTimeline = async (isChecked, layers) => {
     const { setMatchingTimelineCoverage } = this.props;
-    const cmrLayers = await mapGranulesToLayers(layers);
-    const dateRange = this.getNewMatchingDatesRange(cmrLayers);
+    const dateRange = this.getNewMatchingDatesRange(layers);
     setMatchingTimelineCoverage(dateRange, isChecked);
     this.setState({
-      cmrLayers,
       activeLayers: layers,
       shouldIncludeHiddenLayers: isChecked,
     });
@@ -379,12 +285,12 @@ class TimelineLayerCoveragePanel extends Component {
       appNow,
     } = this.props;
     if (layers.length > 0) {
-      return layers.flatMap(({ granules, startDate, endDate }) => {
-        if (!granules?.length) {
+      return layers.flatMap(({ granuleDateRanges, startDate, endDate }) => {
+        if (!granuleDateRanges?.length) {
           return [{ startDate, endDate: endDate || appNow }];
         }
 
-        return granules.map(([start, end]) => ({ startDate: start, endDate: end }));
+        return granuleDateRanges.map(([start, end]) => ({ startDate: start, endDate: end }));
       });
     }
   };
@@ -481,7 +387,6 @@ class TimelineLayerCoveragePanel extends Component {
       timeScale,
     } = this.props;
     const {
-      cmrLayers,
       activeLayers,
       shouldIncludeHiddenLayers,
     } = this.state;
@@ -536,7 +441,7 @@ class TimelineLayerCoveragePanel extends Component {
             </header>
             <Scrollbars style={scrollbarStyle}>
               <CoverageItemList
-                activeLayers={cmrLayers}
+                activeLayers={activeLayers}
                 appNow={appNow}
                 axisWidth={axisWidth}
                 backDate={backDate}
