+++ web/js/modules/layers/constants.js
@@ -21,3 +21,4 @@ export const DEFAULT_NUM_GRANULES = 10;
 export const MIN_GRANULES = 1;
 export const MAX_GRANULES = 30;
 export const UPDATE_DDV_LAYER = 'LAYERS/UPDATE_DDV_LAYER';
+export const ADD_GRANULE_LAYER_GRANULES = 'LAYERS/ADD_GRANULE_LAYER_GRANULES';
