+++ web/js/modules/layers/reducers.js
@@ -22,6 +22,7 @@ import {
   REMOVE_GROUP,
   UPDATE_DDV_LAYER,
   UPDATE_COLLECTION,
+  ADD_GRANULE_LAYER_GRANULES,
 } from './constants';
 import {
   SET_CUSTOM as SET_CUSTOM_PALETTE,
@@ -291,6 +292,23 @@ export function layerReducer(state = initialState, action) {
           },
           granuleFootprints: {
             $set: granuleFootprints,
+          },
+        },
+      });
+    }
+
+    case ADD_GRANULE_LAYER_GRANULES: {
+      const {
+        granuleDateRanges,
+      } = action;
+      const layerIndex = getLayerIndex();
+
+      return update(state, {
+        [compareState]: {
+          layers: {
+            [layerIndex]: {
+              granuleDateRanges: { $set: granuleDateRanges },
+            },
           },
         },
       });
